requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.7
django-money==0.14
django-oauth-toolkit==1.2.0
djangorestframework==3.8.2
flake8==3.5.0
idna==2.7
mccabe==0.6.1
oauthlib==2.1.0
py-moneyed==0.7.0
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
urllib3==1.23
